Changed USer model from Couch to Api
