+++ src/app/app-routing.module.ts
@@ -36,52 +36,42 @@ const routes: Routes = [
           import('./pages/analytics/analytics.module').then((m) => m.AnalyticsModule),
       }, {
         path: 'settings',
-        canActivate: [AuthGuard], data: { roles: ['level3'] },
+        canActivate: [AuthGuard], data: { roles: ['admin'] },
         loadChildren: () =>
           import('./pages/settings/settings.module').then((m) => m.SettingsModule),
       }, {
         path: 'users',
-        canActivate: [AuthGuard], data: { roles: [1, 2, 3] },
+        canActivate: [AuthGuard], data: { roles: ['admin'] },
         loadChildren: () =>
           import('./pages/users/users.module').then((m) => m.UsersModule),
       }, {
         path: 'fields',
-        canActivate: [AuthGuard], data: { roles: ['level3'] },
+        canActivate: [AuthGuard], data: { roles: ['admin'] },
         loadChildren: () =>
           import('./pages/mfields/mfields.module').then((m) => m.MFieldsModule),
       }, {
         path: 'mfield_options',
-        canActivate: [AuthGuard], data: { roles: ['level3'] },
+        canActivate: [AuthGuard], data: { roles: ['admin'] },
         loadChildren: () =>
           import('./pages/mfield_options/mfield_options.module').then((m) => m.MFieldOptionsModule),
       }, {
-        path: 'syndromes',
-        canActivate: [AuthGuard], data: { roles: ['level3'] },
-        loadChildren: () =>
-          import('./pages/syndromes/syndromes.module').then((m) => m.SyndromesModule),
-      }, {
-        path: 'diseases',
-        canActivate: [AuthGuard], data: { roles: ['level3'] },
-        loadChildren: () =>
-          import('./pages/diseases/diseases.module').then((m) => m.DiseasesModule),
-      }, {
         path: 'mforms',
-        canActivate: [AuthGuard], data: { roles: ['level3'] },
+        canActivate: [AuthGuard], data: { roles: ['admin'] },
         loadChildren: () =>
           import('./pages/mforms/mforms.module').then((m) => m.MFormsModule),
       }, {
         path: 'mform_fields',
-        canActivate: [AuthGuard], data: { roles: ['level3'] },
+        canActivate: [AuthGuard], data: { roles: ['admin'] },
         loadChildren: () =>
           import('./pages/mform_fields/mform_fields.module').then((m) => m.MFormFieldsModule),
       }, {
         path: 'mform_forms',
-        canActivate: [AuthGuard], data: { roles: ['level3'] },
+        canActivate: [AuthGuard], data: { roles: ['admin'] },
         loadChildren: () =>
           import('./pages/mform_forms/mform_forms.module').then((m) => m.MFormFormsModule),
       }, {
         path: 'mform_data',
-        canActivate: [AuthGuard], data: { roles: ['level3'] },
+        canActivate: [AuthGuard], data: { roles: ['admin'] },
         loadChildren: () =>
           import('./pages/mform_data/mform_data.module').then((m) => m.MFormDataModule),
       },{
@@ -105,7 +95,7 @@ const routes: Routes = [
       { path: 'assets/:fileName', component: AssetsComponent },
       {
         path: 'thresholds',
-        canActivate: [AuthGuard], data: { roles: ['level3'] },
+        canActivate: [AuthGuard], data: { roles: ['admin'] },
         loadChildren: () =>
           import('./pages/thresholds/thresholds.module').then((m) => m.ThresholdsModule),
       },
