+++ src/app/layouts/full/sidebar/sidebar.component.ts
@@ -3,7 +3,7 @@ import { navItems } from './sidebar-data';
 import { NavService } from '../../../services/nav.service';
 import { AwarenessService } from 'src/app/services/awareness.service';
 import { Router } from '@angular/router';
-import { User } from 'src/app/models/User.model';
+import {User} from "../../../models/User.model";
 
 @Component({
   selector: 'app-sidebar',
@@ -16,17 +16,7 @@ export class SidebarComponent implements OnInit {
   constructor(private router: Router, public navService: NavService, public awareness: AwarenessService) { }
 
   ngOnInit(): void {
-    this.awareness.awaken(() => {
-      this.UserInstance._id = this.awareness.getFocused("authenticated");
 
-      if (this.UserInstance._id != "") {
-        this.UserInstance.acquireInstance((doc: any) => {
-          this.UserInstance.parseInstance(doc);
-        }, (err: any) => {
-          //TODO! Handle errors
-        });
-      }
-    });
   }
 
   onClick(action: any) {
