+++ src/app/pages/authentication/register/register.component.html
@@ -24,7 +24,7 @@
           <mat-label class="mat-subtitle-2 f-s-14 f-w-600 m-b-12 d-block">Name</mat-label>
           <mat-form-field appearance="outline" class="w-100" color="primary">
             <input matInput [formControl]="UserFormControls['user_name']" type="text"
-                   [(ngModel)]="UserInstance.user_name" />
+                   [(ngModel)]="UserInstance.name" />
 
             <mat-error *ngIf="UserFormControls['user_name'].hasError('required')">
               <strong>Required</strong>
@@ -34,7 +34,7 @@
           <mat-label class="mat-subtitle-2 f-s-14 f-w-600 m-b-12 d-block">User Email</mat-label>
           <mat-form-field appearance="outline" class="w-100" color="primary">
             <input matInput [formControl]="UserFormControls['user_email']" type="email"
-                   [(ngModel)]="UserInstance.user_email" />
+                   [(ngModel)]="UserInstance.email" />
 
             <mat-error *ngIf="UserFormControls['user_email'].hasError('required')">
               <strong>Required</strong>
