--- src/app/pages/diseases/composite/composite.component.html
@@ -1,35 +1,0 @@
-<mat-card class="cardWithShadow">
-  <mat-card-header>
-    <mat-card-title>
-      {{ FocusedSyndrome._id != "" ? FocusedSyndrome.syndrome_name : '' }}
-    </mat-card-title>
-    <mat-card-subtitle>
-      <p>Diseases</p>
-    </mat-card-subtitle>
-
-    <span class="flex-1-auto"></span>
-
-    <a mat-flat-button color="primary" class="view-button" [routerLink]="['/diseases/modify']" (click)="setFocused()">
-      Create Disease
-    </a>
-  </mat-card-header>
-
-  <mat-card-content>
-    <mat-card class="b-1 shadow-none view-card" *ngFor="let SeekDisease of CompositeDiseases">
-      <mat-card-header>
-        <mat-card-title>{{ SeekDisease.disease_name }}</mat-card-title>
-      </mat-card-header>
-
-      <mat-card-content class="b-t-1">
-        <p>{{ SeekDisease.disease_description }}</p>
-      </mat-card-content>
-
-      <mat-card-header>
-        <a mat-flat-button color="primary" class="view-button" [routerLink]="['/diseases/modify']"
-          (click)="setFocused(SeekDisease._id)">
-          Edit
-        </a>
-      </mat-card-header>
-    </mat-card>
-  </mat-card-content>
-</mat-card>
