--- src/app/pages/diseases/diseases.module.ts
@@ -1,31 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule } from '@angular/router';
-import { CommonModule } from '@angular/common';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { MaterialModule } from '../../material.module';
-
-// Icons
-import { TablerIconsModule } from 'angular-tabler-icons';
-import * as TablerIcons from 'angular-tabler-icons/icons';
-
-// Components
-import { DiseasesRoutes } from './diseases.routing';
-import { CompositeComponent } from './composite/composite.component';
-import { ModifyComponent } from './modify/modify.component';
-
-@NgModule({
-  imports: [
-    CommonModule,
-    RouterModule.forChild(DiseasesRoutes),
-    MaterialModule,
-    FormsModule,
-    ReactiveFormsModule,
-    TablerIconsModule.pick(TablerIcons)
-  ],
-  declarations: [
-    CompositeComponent,
-    ModifyComponent
-  ]
-})
-export class DiseasesModule {
-}
