--- src/app/pages/diseases/diseases.routing.ts
@@ -1,23 +1,0 @@
-import { Routes } from '@angular/router';
-
-// Pages
-import { CompositeComponent } from './composite/composite.component';
-import { ModifyComponent } from './modify/modify.component';
-
-export const DiseasesRoutes: Routes = [
-  {
-    path: '',
-    children: [
-      {
-        path: '',
-        component: CompositeComponent,
-      }, {
-        path: 'composite',
-        component: CompositeComponent,
-      }, {
-        path: 'modify',
-        component: ModifyComponent,
-      }
-    ],
-  },
-];
