--- src/app/pages/diseases/modify/modify.component.html
@@ -1,46 +1,0 @@
-<mat-card class="cardWithShadow">
-  <mat-card-header>
-    <mat-card-title>
-      <h3>{{DiseaseInstance._action == 'create' ? 'Create' : 'Edit'}} Disease</h3>
-
-      <span class="flex-1-auto"></span>
-    </mat-card-title>
-  </mat-card-header>
-
-  <mat-card-content class="b-t-1">
-    <form class="m-t-30" name="form_user" id="form_user_modify" #ModifyDisease="ngForm" (ngSubmit)="submitInstance()">
-      <mat-label class="mat-subtitle-2 f-s-14 f-w-600 m-b-12 d-block">Syndrome</mat-label>
-      <mat-form-field appearance="outline" class="w-100" color="primary">
-        <input matInput [(ngModel)]="FocusedSyndrome.syndrome_name" readonly />
-      </mat-form-field>
-
-      <mat-label class="mat-subtitle-2 f-s-14 f-w-600 m-b-12 d-block">Name</mat-label>
-      <mat-form-field appearance="outline" class="w-100" color="primary">
-        <input matInput [formControl]="DiseaseFormControls['disease_name']"
-          [(ngModel)]="DiseaseInstance.disease_name" />
-
-        <mat-error *ngIf="DiseaseFormControls['disease_name'].hasError('required')">
-          <strong>Required</strong>
-        </mat-error>
-      </mat-form-field>
-
-      <mat-label class="mat-subtitle-2 f-s-14 f-w-600 m-b-12 d-block">Description</mat-label>
-      <mat-form-field appearance="outline" class="w-100" color="primary">
-        <input matInput [formControl]="DiseaseFormControls['disease_description']"
-          [(ngModel)]="DiseaseInstance.disease_description" />
-      </mat-form-field>
-
-      <div>
-        <button type="submit" mat-flat-button color="primary" class="view-button">
-          {{DiseaseInstance._action == 'create' ? 'Create' : 'Update'}}
-        </button>
-
-        <i-tabler *ngIf="DiseaseInstance._action_result" class="view-icon--inline view-color--green"
-          name="circle-check-filled"></i-tabler>
-
-        <span class="flex-1-auto"></span>
-        <mat-spinner *ngIf="DiseaseInstance._processing" class="view-spinner--inline"></mat-spinner>
-      </div>
-    </form>
-  </mat-card-content>
-</mat-card>
