--- src/app/pages/syndromes/modify/modify.component.html
@@ -1,41 +1,0 @@
-<mat-card class="cardWithShadow">
-  <mat-card-header>
-    <mat-card-title>
-      <h3>{{SyndromeInstance._action == 'create' ? 'Create' : 'Edit'}} Syndrome</h3>
-
-      <span class="flex-1-auto"></span>
-    </mat-card-title>
-  </mat-card-header>
-
-  <mat-card-content class="b-t-1">
-    <form class="m-t-30" name="form_user" id="form_user_modify" #ModifySyndrome="ngForm" (ngSubmit)="submitInstance()">
-      <mat-label class="mat-subtitle-2 f-s-14 f-w-600 m-b-12 d-block">Name</mat-label>
-      <mat-form-field appearance="outline" class="w-100" color="primary">
-        <input matInput [formControl]="SyndromeFormControls['syndrome_name']"
-          [(ngModel)]="SyndromeInstance.syndrome_name" />
-
-        <mat-error *ngIf="SyndromeFormControls['syndrome_name'].hasError('required')">
-          <strong>Required</strong>
-        </mat-error>
-      </mat-form-field>
-
-      <mat-label class="mat-subtitle-2 f-s-14 f-w-600 m-b-12 d-block">Description</mat-label>
-      <mat-form-field appearance="outline" class="w-100" color="primary">
-        <input matInput [formControl]="SyndromeFormControls['syndrome_description']"
-          [(ngModel)]="SyndromeInstance.syndrome_description" />
-      </mat-form-field>
-
-      <div>
-        <button type="submit" mat-flat-button color="primary" class="view-button">
-          {{SyndromeInstance._action == 'create' ? 'Create' : 'Update'}}
-        </button>
-
-        <i-tabler *ngIf="SyndromeInstance._action_result" class="view-icon--inline view-color--green"
-          name="circle-check-filled"></i-tabler>
-
-        <span class="flex-1-auto"></span>
-        <mat-spinner *ngIf="SyndromeInstance._processing" class="view-spinner--inline"></mat-spinner>
-      </div>
-    </form>
-  </mat-card-content>
-</mat-card>
