+++ src/app/pages/users/modify_password/modify_password.component.html
@@ -3,7 +3,7 @@
     <mat-card-header>
       <mat-card-title>
         <h3>
-          {{ UserInstance.user_name }}
+          {{ UserInstance.name }}
         </h3>
       </mat-card-title>
 
@@ -15,7 +15,7 @@
     </mat-card-header>
 
     <mat-card-content>
-      <form class="m-t-30" name="form_user" id="form_user_modify" #ModifyUser="ngForm" (ngSubmit)="submitInstance()">
+      <form class="m-t-30" name="form_user" id="form_user_modify" #ModifyUser="ngForm">
         <!-- Password -->
         <mat-label class="mat-subtitle-2 f-s-14 f-w-600 m-b-12 d-block">Password</mat-label>
         <mat-form-field appearance="outline" class="w-100" color="primary">
@@ -63,10 +63,10 @@
 
           <span class="flex-1-auto"></span>
 
-          <mat-spinner *ngIf="UserInstance.MStatus.ms_processing" class="view-spinner--inline"></mat-spinner>
+<!--          <mat-spinner *ngIf="UserInstance.MStatus.ms_processing" class="view-spinner&#45;&#45;inline"></mat-spinner>-->
 
-          <i-tabler *ngIf="UserInstance.MStatus.ms_action_result" class="view-icon--inline view-color--green"
-                    name="circle-check-filled"></i-tabler>
+<!--          <i-tabler *ngIf="UserInstance.MStatus.ms_action_result" class="view-icon&#45;&#45;inline view-color&#45;&#45;green"-->
+<!--                    name="circle-check-filled"></i-tabler>-->
         </div>
       </form>
     </mat-card-content>
