+++ src/app/pages/users/view/view.component.html
@@ -8,13 +8,13 @@
     <mat-card-header>
       <mat-card-title>
         <h3>
-          {{ UserInstance.user_name }}
+          {{ UserInstance.name }}
         </h3>
       </mat-card-title>
 
       <mat-card-subtitle>
         <p>
-          {{ UserInstance.user_email }}
+          {{ UserInstance.email }}
         </p>
       </mat-card-subtitle>
     </mat-card-header>
