+++ src/app/services/authentication.service.ts
@@ -23,7 +23,7 @@ export class AuthenticationService {
     if (this.awareness.UserInstance._id !== '') {
 
       route_roles.forEach(role => {
-        if (role == this.awareness.UserInstance.user_role) {
+        if (role == this.awareness.UserInstance.role) {
           user_authenticated = true;
         }
       });
@@ -37,8 +37,8 @@ export class AuthenticationService {
   }
 
   getCurrentUserRole(): string {
-    if (this.awareness.UserInstance && this.awareness.UserInstance.user_role) {
-      return this.awareness.UserInstance.user_role;
+    if (this.awareness.UserInstance && this.awareness.UserInstance.role) {
+      return this.awareness.UserInstance.role;
     } else {
       return '';
     }
